requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
croniter==3.0.3
cryptography==43.0.0
discord-py-interactions==5.13.1
discord-typings==0.9.0
emoji==2.12.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4